app/backend/pyproject.toml
[project]
name = "fridge-stats-api"
version = "0.4.0"
description = "FastAPI service exposing fridge sales stats from PostgreSQL (seeded from CSV)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "psycopg2-binary>=2.9"
]
